refactor : import된 api 이름 변경
